Conflicts:
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$10.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$11.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$12.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$13.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$14$1.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$14.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$2$1.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$2.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$3.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$4.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$5.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$6.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$7.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$8.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$9$1.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector$9.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/ChannelSelector.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/MakeChannelSelector.class
	DynamicDisplays/bin/gov/fnal/ppd/dd/display/client/ListOfValidChannels.class
	DynamicDisplays/runSelector.bat
	DynamicDisplays/runSelectorWindow.bat
	DynamicDisplays/src/gov/fnal/ppd/dd/ChannelSelector.java
	DynamicDisplays/versionInformation.dat

+++ DynamicDisplays/src/gov/fnal/ppd/dd/changer/DDButton.java
@@ -252,7 +252,7 @@ public class DDButton extends JButton {
 			String tx = "<html><b>" + channel.getName() + ":</b> " + channel.getDescription() + " <em>(Channel "
 					+ channel.getNumber() + ")</em></html>";
 			if (descriptionFont == null) {
-				float siz = ((float) userDefinedFont.getSize()) / (tx.length() / 60.0f);
+				float siz = (userDefinedFont.getSize()) / (tx.length() / 60.0f);
 				if (siz < 7.0f)
 					siz = 7.0f;
 				else if (siz > userDefinedFont.getSize())
@@ -273,8 +273,7 @@ public class DDButton extends JButton {
 					tx = tx.replace("/", " /");
 			}
 			if (urlFont == null) {
-				float siz = fixFontSize(((float) userDefinedFont.getSize()) / (tx.length() / 25.0f), 7.0f,
-						(float) userDefinedFont.getSize());
+				float siz = fixFontSize((userDefinedFont.getSize()) / (tx.length() / 25.0f), 7.0f, userDefinedFont.getSize());
 				urlFont = userDefinedFont.deriveFont(siz);
 			}
 			setText("<html>" + tx + "</html>");
